requirements.txt
Flask==2.2.5
openai==1.30.5
python-dotenv==0.21.0